Louvain's method for bipartite graphs (draft)
The Louvain's method is a graph hierarchical clustering algorithm by
optimization of a modularity function. It was first described in 2008 by
Blondel et al. using Newman's modularity. This method do not guarantee
to reach the global optimum, but happens to be a powerful heuristic to
treat very large graphs. In the following, we propose a derivation of
this method for Guimera's bipartite modularity function. Moreover we
propose a first implementation of this algorithm for the =rgraph=
library.

Louvain's method is a two-step iterative process: in the first step,
the algorithm loop through each individual vertex, putting it in the
group of its neighbor that give the best modularity increase. The
looping continue until further increase in modularity is impossible
(or below a threshold). Note that in this case, the neighboring vertices
considered are the ones that share a team with the focal
agent (i.e. neighbors vertices in the projected graph of the agents).
In the second step (non implemented) a new graph is built using the groups found
previously as vertices.