Replace tdestroy() calls: lib/ not needed anymore.
This commit:
- Reimplements FreeLabelDict without using tdestroy().
- Replaces stray calls to tdestroy function by FreeLabelDict().
- Reduces the number of parameter of FreeNodeTree().

The issue:

rgraph uses binary search tree when it need to access to a node by
label. For that it relies on the <search.h> header file that define
the tsearch() function, used to build and access binary seach tree, and
part of the POSIX standard.

However, to recursively free the memory of the tree created, rgraph used
to use tdestroy(), a GNU extension of the <search.h> library that is not
covered by the POSIX standard, and thus absent on some system (Windows,
even with minGW, recent MACOS...).

To be able to compile on those platform, part of GNUlibc was provided in
the lib/ folder. However, this required a call to gnulib and autoconf
to generate the proper header, which is not practical if we want the
library to be compiled as a part of a R-package.

Thus, I reimplemented the FreeLabelDict() function to avoid using
tdestroy(). It is fairly simple and should not leak memory. This commit
remove the need for the lib/ directory on POSIX-compatible systems.