Weighted Louvain's method.
The weighted version only change the defintion of the preprocessed
values (that are used to compute the modularity gain).

It means that it uses:
- NodeStrengthFast rather CountLinks
- SumProductsOfCommonWeightsBipart rather than CountLinks

Also, epsilon, in the deniminator of the first component of the
modularity (\sum\limits_a m_a(m_a-\epsilon)}) is set to the minimum
weight of the edges of a (and not 1 as in the SA version !). You could
argue to use the average or the median for a different approximation.