Getting all SA-related functions in one file.
We have 4 cases to cover:
- Single mode graphs,
- Single mode weighted graphs,
- Bipartite graphs,
- Bipartite weighted graphs.

As of today, those 4 cases have each their SACommunityIdent and
SAGroupSplit functions, which sum to ~ 2000 lines of codes doing
essentially the same thing four time.

My goal is to understand those function, and refactor them to remove
redundancies and speed up things (hopefully with a little bit of
OpenMP). This commit starts this journey by grouping this 8 functions
in a sannealing.c file.

Some notes:

The structure of the 4 SACommunityIdent is more or less the same:
1) Init, place nodes in groups.
2) Temperature loop.
   2a) Individual movments.
   2b) Global movments: first merge, then split if they are empty sets.
3) Free memory.

Each movement (individual, merge, split) is composed of:
1) Propose movement (for split it uses the SAGroupSplit function)
2) Compute change in energy
3) Accept or not according to the SA criterion ('Metroppolis')

The SAGroupSplit function uses a small SA (with only individual
movements) to divide the group into two. (It can also take the
components if the group in unconnected).