Saner Metropolis-Boltzman conditions for module splitting.
In the old implementation it use to be:

merge back if: (dE>0) && (rdm()>exp(-dE/T))

which is equivalent to:
merge back if NOT: (-dE>=0) || (rdm()<=exp(-dE/T))

Using || rather than && enable lasy evaluation: if -dE>=0, the random
number is NOT generated.

Also instead of -dE, I used dE so it is consistant with the others lines
testing for this criterion, and the minus sign is applied directly when
computing the energy gain (dE = -dEMergeModules()).

Finally, to avoid useless modifications of the global energy when the
split is trivial (one of the new modules is empty), I added a condition
to skip this block entirely depending on the size of the splitted
modules.

Otherwise it could mess with the nochange count (as the
modularity change slightly because we actually move stuff around, even
if it is trivial). I think that this was the reason for the
weird (dE>EPSILON_MOD_B) in the legacy bipartite code. This is much
cleaner IMHO.