Draft to have a way to load partition from file and only compute the modularity and the roles.
It can be use with the command netcarto -f network.ntw -p
partition.part. The input/output function will probably ends up moving
out of this file.

The building of the partition from the file uses a hash table (via the
posix implementation) rather than the binary trees that are used in
other parts of the library.